+++ src/util.h
@@ -165,7 +165,7 @@ public:
 
 bool getNewsRSS(std::vector<WebsiteNews>& WebsiteNewsList);
 enum yahoo_price_type { FIAT = 0, SHARES };
-bool getOnlineCurrencyRates(wxString& msg, const int curr_id = -1, const bool used_only = true);
+bool getOnlineCurrencyRates(wxString& msg, const int64 curr_id = -1, const bool used_only = true);
 bool get_yahoo_prices(std::map<wxString, double>& symbols
     , std::map<wxString, double>& out
     , const wxString& base_currency_symbol
